+++ CHANGELOG.md
@@ -3,7 +3,10 @@
 ## Release_1.3.0_20251121_build_A
 
 ### 功能构建
-  
+
+- 实现运维指令。
+  - com.dwarfeng.statistics.impl.service.telqos.SupportCommand。
+
 - 优化支持实体机制。
   - 新建支持 QoS 服务 com.dwarfeng.statistics.stack.service.SupportQosService。
   - 将支持实体维护服务的重置功能迁移至 QoS 服务。